sysfs: stubs in approach to defining a new file type
After the last couple months of thinking, I eventually got to the idea
that defining a new file type has two problems:

* What sort of type do we define?
* How in the world do we migrate to a new type?

What this PR does is spike an approach to the latter without solidifing
the former. My thinking is that the best way is to incrementally migrate
functions one-by-one to a new type. Meanwhile, temporarily expose a
`.File()` function to access the `fs.File` we want to rid eventually.
This makes it easy to identify what's left as you just look for any
calls to `.File()`.

If this approach helps, lemme know and I can polish for merge. I mainly
had this branch locally so want to decide to delete it or not!

Signed-off-by: Adrian Cole <adrian@tetrate.io>